Revert "IMPALA-1619: Support 64-bit allocations."
This reverts commit 1ffb2bd5a2a2faaa759ebdbaf49bf00aa8f86b5e.

Unbreak the packaging builds for now.

Reviewed-on: http://gerrit.cloudera.org:8080/3543
Tested-by: Michael Ho <kwho@cloudera.com>

+++ be/src/runtime/collection-value-builder.h
@@ -32,7 +32,7 @@ class CollectionValueBuilder {
 
   CollectionValueBuilder(CollectionValue* coll_value, const TupleDescriptor& tuple_desc,
       MemPool* pool, RuntimeState* state,
-      int64_t initial_tuple_capacity = DEFAULT_INITIAL_TUPLE_CAPACITY)
+      int initial_tuple_capacity = DEFAULT_INITIAL_TUPLE_CAPACITY)
     : coll_value_(coll_value),
       tuple_desc_(tuple_desc),
       pool_(pool),
